RECMSNEW/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-cors-headers==2.3.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.4
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
urllib3==1.22
dnspython3==1.15.0
django-filter==2.0.0
